refactor: Move rd outside rmi module
Realm Descriptor (RD) has nothing to do with Realm Management
Interface (RMI). Thus, it would be better to locate rd module
outside rmi module. Note that params has not been moved, as it
represents `RmiRealmParams` type.

[before]
rmm/src/rmi/realm/rd.rs

[after]
rmm/src/realm/rd.rs
