+++ rmm/src/gic.rs
@@ -1,7 +1,7 @@
+use crate::realm::rd::Rd;
 use crate::realm::vcpu::VCPU;
 use crate::rmi::error::Error;
 use crate::rmi::error::InternalError::*;
-use crate::rmi::realm::Rd;
 use crate::rmi::rec::run::Run;
 
 use armv9a::regs::*;
