+++ rmm/src/mmio.rs
@@ -1,6 +1,6 @@
+use crate::realm::rd::Rd;
 use crate::rmi::error::Error;
 use crate::rmi::error::InternalError::*;
-use crate::rmi::realm::Rd;
 use crate::rmi::rec::run::{Run, REC_ENTRY_FLAG_EMUL_MMIO};
 use armv9a::regs::*;
