+++ rmm/src/realm/mm/stage2_tte.rs
@@ -4,8 +4,8 @@ use vmsa::address::PhysAddr;
 use super::address::GuestPhysAddr;
 use crate::granule::GRANULE_SIZE;
 use crate::realm::mm::page_table::pte::{attribute, shareable};
+use crate::realm::rd::Rd;
 use crate::rmi::error::Error;
-use crate::rmi::realm::Rd;
 use crate::rmi::rtt::{RTT_MIN_BLOCK_LEVEL, RTT_PAGE_LEVEL};
 use armv9a::{define_bitfield, define_bits, define_mask};
 use vmsa::guard::Content;
