+++ rmm/src/realm/mod.rs
@@ -1,6 +1,7 @@
 pub mod config;
 pub mod context;
 pub mod mm;
+pub mod rd;
 pub mod registry;
 pub mod timer;
 pub mod vcpu;
