+++ rmm/src/rmi/rec/exit.rs
@@ -6,12 +6,12 @@ use crate::granule::GranuleState;
 use crate::granule::GRANULE_MASK;
 use crate::realm::context::get_reg;
 use crate::realm::mm::stage2_tte::S2TTE;
+use crate::realm::rd::Rd;
 use crate::rmi::error::Error;
 use crate::rmi::rec::run::Run;
 use crate::rmi::rec::Rec;
 use crate::rmi::rtt::is_protected_ipa;
 use crate::rmi::rtt::RTT_PAGE_LEVEL;
-use crate::rmi::Rd;
 use crate::Monitor;
 use crate::{rmi, rsi};
 use armv9a::{EsrEl2, EMULATABLE_ABORT_MASK, HPFAR_EL2, NON_EMULATABLE_ABORT_MASK};
