+++ rmm/src/rmi/rec/handlers.rs
@@ -12,10 +12,10 @@ use crate::host;
 use crate::listen;
 use crate::measurement::HashContext;
 use crate::realm::context::set_reg;
+use crate::realm::rd::{Rd, State};
 use crate::realm::vcpu::create_vcpu;
 use crate::rmi;
 use crate::rmi::error::Error;
-use crate::rmi::realm::{rd::State, Rd};
 use crate::rmi::rec::exit::handle_realm_exit;
 use crate::rmi::rec::RecState;
 use crate::rsi::do_host_call;
