+++ rmm/src/rmi/rec/vtcr.rs
@@ -1,6 +1,6 @@
 use super::Rec;
+use crate::realm::rd::Rd;
 use crate::rmi::error::Error;
-use crate::rmi::realm::Rd;
 use armv9a::bits_in_reg;
 use armv9a::regs::*;
