+++ rmm/src/rsi/mod.rs
@@ -14,9 +14,9 @@ use crate::realm::config::realm_config;
 use crate::realm::context::{get_reg, set_reg};
 use crate::realm::mm::address::GuestPhysAddr;
 use crate::realm::mm::stage2_tte::invalid_ripas;
+use crate::realm::rd::Rd;
 use crate::rmi;
 use crate::rmi::error::Error;
-use crate::rmi::realm::Rd;
 use crate::rmi::rec::run::Run;
 use crate::rmi::rec::{Rec, RmmRecAttestState};
 use crate::rmi::rtt::{is_protected_ipa, validate_ipa, RTT_PAGE_LEVEL};
