+++ rmm/src/rsi/psci.rs
@@ -2,8 +2,8 @@ use crate::event::RsiHandle;
 use crate::granule::GranuleState;
 use crate::listen;
 use crate::realm::context::{get_reg, set_reg};
+use crate::realm::rd::{Rd, State};
 use crate::rmi;
-use crate::rmi::realm::{rd::State, Rd};
 use crate::rmi::rec::run::Run;
 use crate::rmi::rec::Rec;
 use crate::Monitor;
