+++ rmm/src/rtt.rs
@@ -6,8 +6,8 @@ use crate::realm::mm::page_table::pte::attribute;
 use crate::realm::mm::page_table::pte::{permission, shareable};
 use crate::realm::mm::stage2_tte::{desc_type, invalid_hipas, invalid_ripas};
 use crate::realm::mm::stage2_tte::{RttPage, INVALID_UNPROTECTED, S2TTE};
+use crate::realm::rd::Rd;
 use crate::rmi::error::Error;
-use crate::rmi::realm::Rd;
 use crate::rmi::rtt::RTT_PAGE_LEVEL;
 use crate::rmi::rtt::S2TTE_STRIDE;
 use crate::rmi::rtt_entry_state;
